requirements.txt
moviepy==1.0.3
edge-tts>=6.1.9
requests>=2.31.0
pillow>=10.0.0
rich>=13.7.0
python-dotenv>=1.0.0
g4f>=0.2.0
asyncio
streamlit
watchdog
